+++ backend/routes/search.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const { search } = require("../controllers/searchController");
+
+const router = express.Router();
+
+router.get("/search", search);
+
+module.exports = router;
