+++ frontend/src/components/Map.js
@@ -2,6 +2,7 @@ import React, { useRef, useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import SearchPanel from "./SearchPanel.js";
 import loadGoogleMapsScript from "./loadGoogleMap";
+import NavigationComponent from "./NavigationComponent";
 
 const fetchGeoJSON = async (url) => {
   try {
@@ -22,6 +23,8 @@ const MapComponent = () => {
   const [currentFloor, setCurrentFloor] = useState(null);
   const [buildingFloors, setBuildingFloors] = useState([]);
   const [highlightedRoom, setHighlightedRoom] = useState(null);
+  const [startLocation, setStartLocation] = useState(null);
+  const [endLocation, setEndLocation] = useState(null);
   const [mapZoom, setMapZoom] = useState(18);
   const [mapCenter, setMapCenter] = useState({ lat: 47.693344, lng: 17.627529 });
   const [loading, setLoading] = useState(true);
@@ -266,6 +269,8 @@ const MapComponent = () => {
       const data = await response.json();
   
       console.log("Keresési eredmények:", data);
+
+      
   
       if (data.buildings.length > 0) {
         setIsBuildingView(false); // Külső nézetre váltás
@@ -410,14 +415,108 @@ const MapComponent = () => {
     });
   };
 
+  const handleNavigate = (from, to) => {
+    setStartLocation(from);
+    setEndLocation(to);
+  };
+
+  const handleRouteSearch = async (start, destination) => {
+    if (!start || !destination) {
+      alert("Mindkét helyet meg kell adnod az útvonaltervezéshez!");
+      return;
+    }
+  
+    try {
+      // Keresés a kezdő- és végpont szobához
+      const responseStart = await fetch(`http://localhost:5000/api/search?q=${start}`);
+      const responseEnd = await fetch(`http://localhost:5000/api/search?q=${destination}`);
+  
+      const dataStart = await responseStart.json();
+      const dataEnd = await responseEnd.json();
+  
+      if (dataStart.rooms.length > 0 && dataEnd.rooms.length > 0) {
+        const startRoom = dataStart.rooms[0];
+        const endRoom = dataEnd.rooms[0];
+
+        // Beltéri vagy épületközi útvonal?
+        if (startRoom.floor.building.id === endRoom.floor.building.id) {
+          // Beltéri útvonal keresése
+          const pathResponse = await fetch(`http://localhost:5000/api/path?fromRoom=${startRoom.id}&toRoom=${endRoom.id}`);
+          const pathData = await pathResponse.json();
+  
+          if (pathData.waypoints) {
+            drawPathOnMap(pathData.waypoints);
+          } else {
+            alert("Nincs elérhető beltéri útvonal!");
+          }
+        } else {
+          // Két épület közötti útvonal keresése
+          const connectionResponse = await fetch(
+            `http://localhost:5000/api/connection?fromBuilding=${startRoom.floor.building.id}&toBuilding=${endRoom.floor.building.id}&fromFloor=${startRoom.floor.number}&toFloor=${endRoom.floor.number}`
+          );
+          const connectionData = await connectionResponse.json();
+  
+          if (connectionData.waypoints) {
+            drawPathOnMap(connectionData.waypoints);
+          } else {
+            alert("Nincs kapcsolat az épületek között!");
+          }
+        }
+
+        // Épület nézet aktiválása és szobák kiemelése
+        setIsBuildingView(true);
+        highlightRoom(startRoom);
+        highlightRoom(endRoom);
+  
+        // Útvonal kezdő- és végpontját térképre állítjuk
+        handleNavigate(
+          { lat: JSON.parse(startRoom.coordinates)[0][1], lng: JSON.parse(startRoom.coordinates)[0][0] },
+          { lat: JSON.parse(endRoom.coordinates)[0][1], lng: JSON.parse(endRoom.coordinates)[0][0] }
+        );
+  
+      } else {
+        alert("Nincs elegendő találat az útvonaltervezéshez!");
+      }
+    } catch (error) {
+      console.error("🚨 Hiba az útvonaltervezés során:", error);
+    }
+  };
+
+
+  const drawPathOnMap = (waypoints) => {
+    if (!map.current) return;
+  
+    // Előző útvonal törlése
+    if (window.currentPath) {
+      window.currentPath.setMap(null);
+    }
+  
+    const pathCoordinates = waypoints.map(([lng, lat]) => ({ lat, lng }));
+  
+    window.currentPath = new window.google.maps.Polyline({
+      path: pathCoordinates,
+      geodesic: true,
+      strokeColor: "#FF0000",
+      strokeOpacity: 1.0,
+      strokeWeight: 3,
+    });
+  
+    window.currentPath.setMap(map.current);
+  };
+  
+  
+  
+
   return (
     <div style={{ position: "relative", width: "100%", height: "100vh" }}>
       <SearchPanel
         onSearch={handleSearch}
         highlightBuilding={highlightBuilding}
         highlightRoom={highlightRoom}
+        onRouteSearch={handleRouteSearch}
         onGroupSelect={handleGroupSelect}
       />
+      <NavigationComponent start={startLocation} end={endLocation} map={map.current} />
       {loading && <p>Betöltés...</p>}
       {error && <p style={{ color: "red" }}>Hiba történt: {error}</p>}
       <div ref={mapContainer} style={{ width: "100%", height: "100%" }} />
